+++ packages/native-with-source/src/ByteReaderWithPositionDriver.cpp
@@ -54,7 +54,7 @@ napi_value ByteReaderWithPositionDriver::Read(const CallbackArgs &args) {
   IOBasicTypes::LongBufferSizeType size = ToUint32(args.Env(), args[0]);
   std::vector<IOBasicTypes::Byte> buffer(size);
   size = driver->mInstance->Read(buffer.data(), static_cast<int>(size));
-  return BytesToArray(args.Env(), buffer.data(), size);
+  return BytesToBuffer(args.Env(), buffer.data(), size);
 }
 
 napi_value ByteReaderWithPositionDriver::NotEnded(const CallbackArgs &args) {